Add support for Lazer file format.

+++ Mapping_Tools/Classes/BeatmapHelper/Events/IHasEndTime.cs
@@ -3,6 +3,6 @@
     /// Indicates that a type has an end time. Used by Property Transformer on Events
     /// </summary>
     public interface IHasEndTime {
-        int EndTime { get; set; }
+        double EndTime { get; set; }
     }
 }
