+++ Mapping_Tools/Classes/BeatmapHelper/Events/IHasStartTime.cs
@@ -3,6 +3,6 @@
     /// Indicates that a type has a start time. Used by Property Transformer on Events
     /// </summary>
     public interface IHasStartTime {
-        int StartTime { get; set; }
+        double StartTime { get; set; }
     }
 }
